Move AgentClassLoader out of public API.
