+++ auto-api/src/test/groovy/io/opentelemetry/instrumentation/auto/api/InternalJarURLHandlerTest.groovy
@@ -17,6 +17,7 @@
 package io.opentelemetry.instrumentation.auto.api
 
 import io.opentelemetry.auto.util.test.AgentSpecification
+import io.opentelemetry.javaagent.bootstrap.InternalJarURLHandler
 import spock.lang.Shared
 
 class InternalJarURLHandlerTest extends AgentSpecification {
