+++ javaagent-bootstrap/javaagent-bootstrap.gradle
@@ -15,7 +15,6 @@ dependencies {
   // ^ Generally a bad idea for libraries, but we're shadowing.
 
   implementation project(':instrumentation-api')
-  implementation project(':auto-api')
 
   testImplementation project(':testing-common')
   testImplementation group: 'org.mockito', name: 'mockito-core', version: '2.19.0'
