+++ javaagent-tooling/src/test/groovy/io/opentelemetry/javaagent/test/ClassLoaderMatcherTest.groovy
@@ -17,7 +17,7 @@
 package io.opentelemetry.auto.test
 
 import io.opentelemetry.auto.util.test.AgentSpecification
-import io.opentelemetry.instrumentation.auto.api.AgentClassLoader
+import io.opentelemetry.javaagent.bootstrap.AgentClassLoader
 import io.opentelemetry.javaagent.tooling.ClassLoaderMatcher
 import io.opentelemetry.javaagent.tooling.ExporterClassLoader
 import io.opentelemetry.javaagent.tooling.log.LogContextScopeListener
